requirements.txt
# ============================================================================
# AadhaarTrust - Complete Requirements
# ============================================================================
# This file includes all dependencies for both the existing pipeline
# and the new FastAPI backend API
# ============================================================================

# ============================================================================
# Backend API Dependencies (FastAPI)
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================================================
# Existing Pipeline Dependencies
# ============================================================================

# Computer Vision & Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0

# Deep Learning & ML 
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# QR Code Processing
zxing-cpp>=2.2.0

# RAW Image Processing
rawpy>=0.5.0

# ============================================================================
# Optional Dependencies (for future enhancements)
# ============================================================================

# OCR (uncomment when implementing OCR)
# python-doctr[torch]>=0.7.0
# pytesseract>=0.3.10

# PDF Generation (uncomment when implementing reports)
# reportlab>=4.0.0
# weasyprint>=60.0

# Additional utilities
# requests>=2.31.0  # For external API calls if needed
# aiofiles>=23.0.0  # For async file operations

# ============================================================================
# Development Dependencies (optional)
# ============================================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
